Include readme in sponsors change
Whenever the sponsors.md changes, the include in the readme should also be changed. Rather than waiting for a second PR to be created for that, merge both actions in a single workflow.

We don't need the includes workflow anymore, therefore.

Force run on workflow

Due to the CRLLF/LF issue, the workflow is otherwise failing on run.

See https://github.com/peter-evans/create-pull-request/issues/1202